refactor(notifications): refactor notifications page header to its own file
